chore(app): rename context
