+++ src/main.jsx
@@ -1,7 +1,7 @@
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import App from './components/app/App.jsx';
-import { AuthProvider } from './components/contexts/UserContext.jsx';
+import { AuthProvider } from './components/contexts/AuthContext.jsx';
 import '@arcgis/core/assets/esri/themes/light/main.css';
 import './index.css';
